move allow_list_type to config_file
