+++ src/config/transactionAllowList.ts
@@ -1,4 +1,17 @@
-import { TransactionAllow } from 'src/shared/entities';
+export interface ComparisonOperation {
+  eq?: string; // txValue == condition is allowed
+  nq?: string; // txValue != condition is allowed
+  gt?: string; // txValue > condition is allowed
+  gte?: string; // txValue >= condition is allowed
+  lt?: string; // txValue < condition is allowed
+  lte?: string; // txValue <= condition is allowed
+}
+
+export interface TransactionAllow {
+  fromList: Array<string>;
+  toList: Array<string>;
+  value?: ComparisonOperation;
+}
 
 export const getTxAllowList = (): Array<TransactionAllow> => {
   return [
