--- src/shared/entities/TransactionAllow.ts
@@ -1,14 +1,0 @@
-export interface ComparisonOperation {
-  eq?: string; // txValue == condition is allowed
-  nq?: string; // txValue != condition is allowed
-  gt?: string; // txValue > condition is allowed
-  gte?: string; // txValue >= condition is allowed
-  lt?: string; // txValue < condition is allowed
-  lte?: string; // txValue <= condition is allowed
-}
-
-export interface TransactionAllow {
-  fromList: Array<string>;
-  toList: Array<string>;
-  value?: ComparisonOperation;
-}
