+++ src/shared/entities/index.ts
@@ -1,4 +1,3 @@
-export * from './TransactionAllow';
 export * from './TransactionParam';
 export * from './Jsonrpc';
 export * from './Verse';
